move quick_color_check to qc submodule
