+++ mkdocs.yml
@@ -145,6 +145,7 @@ nav:
       - 'Plot Image': plot_image.md
       - "Quality Control Tools":
         - "Exposure": qc_exposure.md
+        - "Quick Color Check": quick_color_check.md
       - 'Read Image': read_image.md
       - 'Read Bayer Raw Image': read_bayer.md
       - 'Report Size Marker': report_size_marker.md
@@ -187,7 +188,6 @@ nav:
           - 'Create Color Card Mask': create_color_card_mask.md
           - 'Convert Color Card to Matrix': get_color_matrix.md
           - 'Color Correction Workflow': transform_correct_color.md
-          - 'Quick Color Check': quick_color_check.md
           - 'Affine Color Correction': transform_affine_color_correction.md
           - 'Standard Color Matrix': std_color_matrix.md
           - 'Gamma Correction': transform_gamma_correct.md
