+++ plantcv/plantcv/qc/__init__.py
@@ -1,3 +1,5 @@
 from plantcv.plantcv.qc.exposure import exposure
+from plantcv.plantcv.qc.quick_color_check import quick_color_check
 
-__all__ = ["exposure"]
+
+__all__ = ["exposure", "quick_color_check"]
