+++ plantcv/plantcv/transform/__init__.py
@@ -6,7 +6,6 @@ from plantcv.plantcv.transform.color_correction import save_matrix
 from plantcv.plantcv.transform.color_correction import load_matrix
 from plantcv.plantcv.transform.color_correction import correct_color
 from plantcv.plantcv.transform.color_correction import create_color_card_mask
-from plantcv.plantcv.transform.color_correction import quick_color_check
 from plantcv.plantcv.transform.color_correction import std_color_matrix
 from plantcv.plantcv.transform.color_correction import astro_color_matrix
 from plantcv.plantcv.transform.color_correction import affine_color_correction
@@ -24,7 +23,7 @@ from plantcv.plantcv.transform.auto_correct_color import auto_correct_color
 from plantcv.plantcv.transform.auto_correct_color import auto_correct_color_nonlinear
 
 __all__ = ["get_color_matrix", "get_matrix_m", "calc_transformation_matrix", "apply_transformation_matrix",
-           "save_matrix", "load_matrix", "correct_color", "create_color_card_mask", "quick_color_check",
+           "save_matrix", "load_matrix", "correct_color", "create_color_card_mask",
            "std_color_matrix", "affine_color_correction", "rescale", "nonuniform_illumination", "resize",
            "resize_factor", "warp", "rotate", "warp", "warp_align", "gamma_correct", "detect_color_card", "checkerboard_calib",
            "calibrate_camera", "merge_images", "auto_correct_color", "mask_color_card", "auto_correct_color_nonlinear",
